restructure druglab.db.backend module
